Moving updater package under storage as it is only being used there
