+++ pkg/storage/export/export/BUILD.bazel
@@ -19,9 +19,9 @@ go_library(
         "//pkg/controller:go_default_library",
         "//pkg/instancetype:go_default_library",
         "//pkg/storage/snapshot:go_default_library",
+        "//pkg/storage/status:go_default_library",
         "//pkg/storage/types:go_default_library",
         "//pkg/util:go_default_library",
-        "//pkg/util/status:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-controller/watch/util:go_default_library",
